revert trimming and this somehow fixes the antibot issues??????
idk but it works somehow??????

+++ inc/anti-bot.php
@@ -10,7 +10,7 @@ $hidden_inputs_twig = array();
 
 class AntiBot {
 	public $salt, $inputs = array(), $index = 0;
-
+	
 	public static function randomString($length, $uppercase = false, $special_chars = false, $unicode_chars = false) {
 		$chars = 'abcdefghijklmnopqrstuvwxyz0123456789';
 		if ($uppercase)
@@ -22,11 +22,11 @@ class AntiBot {
 			for ($n = 0; $n < $len; $n++)
 				$chars .= mb_convert_encoding('&#' . mt_rand(0x2600, 0x26FF) . ';', 'UTF-8', 'HTML-ENTITIES');
 		}
-
+		
 		$chars = preg_split('//u', $chars, -1, PREG_SPLIT_NO_EMPTY);
-
+		
 		$ch = array();
-
+		
 		// fill up $ch until we reach $length
 		while (count($ch) < $length) {
 			$n = $length - count($ch);
@@ -39,40 +39,40 @@ class AntiBot {
 			foreach ($keys as $key)
 				$ch[] = $chars[$key];
 		}
-
+		
 		$chars = $ch;
-
+		
 		return implode('', $chars);
 	}
-
+	
 	public static function make_confusing($string) {
 		$chars = preg_split('//u', $string, -1, PREG_SPLIT_NO_EMPTY);
-
+		
 		foreach ($chars as &$c) {
 			if (mt_rand(0, 3) != 0)
 				$c = utf8tohtml($c);
 			else
 				$c = mb_encode_numericentity($c, array(0, 0xffff, 0, 0xffff), 'UTF-8');
 		}
-
+		
 		return implode('', $chars);
 	}
-
+	
 	public function __construct(array $salt = array()) {
 		global $config;
-
+		
 		if (!empty($salt)) {
 			// create a salted hash of the "extra salt"
 			$this->salt = implode(':', $salt);
-		} else {
+		} else { 
 			$this->salt = '';
 		}
-
+		
 		shuffle($config['spam']['hidden_input_names']);
-
+		
 		$input_count = mt_rand($config['spam']['hidden_inputs_min'], $config['spam']['hidden_inputs_max']);
 		$hidden_input_names_x = 0;
-
+		
 		for ($x = 0; $x < $input_count ; $x++) {
 			if ($hidden_input_names_x === false || mt_rand(0, 2) == 0) {
 				// Use an obscure name
@@ -83,7 +83,7 @@ class AntiBot {
 				if ($hidden_input_names_x >= count($config['spam']['hidden_input_names']))
 					$hidden_input_names_x = false;
 			}
-
+			
 			if (mt_rand(0, 2) == 0) {
 				// Value must be null
 				$this->inputs[$name] = '';
@@ -96,16 +96,16 @@ class AntiBot {
 			}
 		}
 	}
-
+	
 	public static function space() {
 		if (mt_rand(0, 3) != 0)
 			return ' ';
 		return str_repeat(' ', mt_rand(1, 3));
 	}
-
+	
 	public function html($count = false) {
 		global $config;
-
+		
 		$elements = array(
 			'<input type="hidden" name="%name%" value="%value%">',
 			'<input type="hidden" value="%value%" name="%name%">',
@@ -119,13 +119,13 @@ class AntiBot {
 			'<textarea style="display:none" name="%name%">%value%</textarea>',
 			'<textarea name="%name%" style="display:none">%value%</textarea>'
 		);
-
+		
 		$html = '';
-
+		
 		if ($count === false) {
-			$count = mt_rand(1, (int)abs(count($this->inputs) / 15) + 1);
+			$count = mt_rand(1, abs(count($this->inputs) / 15) + 1);
 		}
-
+		
 		if ($count === true) {
 			// all elements
 			$inputs = array_slice($this->inputs, $this->index);
@@ -133,7 +133,7 @@ class AntiBot {
 			$inputs = array_slice($this->inputs, $this->index, $count);
 		}
 		$this->index += count($inputs);
-
+		
 		foreach ($inputs as $name => $value) {
 			$element = false;
 			while (!$element) {
@@ -146,37 +146,37 @@ class AntiBot {
 					$element = false;
 				}
 			}
-
+			
 			$element = str_replace('%name%', utf8tohtml($name), $element);
-
+			
 			if (mt_rand(0, 2) == 0)
 				$value = $this->make_confusing($value);
 			else
 				$value = utf8tohtml($value);
-
+			
 			if (strpos($element, 'textarea') === false)
 				$value = str_replace('"', '&quot;', $value);
-
+			
 			$element = str_replace('%value%', $value, $element);
-
+			
 			$html .= $element;
 		}
-
+		
 		return $html;
 	}
-
+	
 	public function reset() {
 		$this->index = 0;
 	}
-
+	
 	public function hash() {
 		global $config;
-
+		
 		// This is the tricky part: create a hash to validate it after
 		// First, sort the keys in alphabetical order (A-Z)
 		$inputs = $this->inputs;
 		ksort($inputs);
-
+		
 		$hash = '';
 		// Iterate through each input
 		foreach ($inputs as $name => $value) {
@@ -184,7 +184,7 @@ class AntiBot {
 		}
 		// Add a salt to the hash
 		$hash .= $config['cookies']['salt'];
-
+		
 		// Use SHA1 for the hash
 		return sha1($hash . $this->salt);
 	}
